Refactored upload component to video module.
